Update site URL in admin config
The `site_url` in the admin configuration was updated to reflect the current website this CMS admin is managing, changing from `ntn-boilerplate.netlify.app` to `abbottdances.org`. This change ensures that all CMS content routing and relative path resolutions will work correctly for the current site. This update is critical to ensure that the CMS works for the right domain. No other parts of the system are impacted by this change.